TUC-128 Better handling of target units
The Target now has a current unit and a unit for the final result.
It allows to convert the target evaluation to the correct unit at the end
of target evaluation.

Also exception handling in ComputingComponent, so if a calculation went badly,
the exception is catched and Result::Error is returned instead of crashing
the software.

	modified:   src/tucucore/computingcomponent.cpp
		* Added exception handling
	modified:   src/tucucore/drugtreatment/target.cpp
		* Added a unit and final unit field
	modified:   src/tucucore/drugtreatment/target.h
		* Added a unit and final unit field
	modified:   src/tucucore/parameter.h
		* Something related to other work, and equal function
	modified:   src/tucucore/targetevaluationresult.cpp
		* Added a unit in the target evaluation result
	modified:   src/tucucore/targetevaluationresult.h
		* Added a unit in the target evaluation result
	modified:   src/tucucore/targetevaluator.cpp
		* The result is converted to the correct unit
	modified:   src/tucucore/targetextractor.cpp
		* The construction of Target takes care of the unit