TUC-150 Implement ResidualErrorModelExtractor
Split of the previous ResidualErrorModel.
Now the DrugModel analytes have an ErrorModel object, and it is translated
into an IResidualErrorModel by a ResidualErrorModelExtractor in the
ComputingComponent.

For now the translation is direct, but it could use the covariates to
calculate the sigmas. Be careful though that having different sigmas could
be manageable for likelihood calculation, but would be difficult for applying
etas for percentiles calculations.

	modified:   qtcreator/tucucore.pri
	modified:   ../src/tucucore/computingcomponent.cpp
	modified:   ../src/tucucore/drugmodel/analyte.h
	new file:   ../src/tucucore/drugmodel/errormodel.cpp
	new file:   ../src/tucucore/drugmodel/errormodel.h
	modified:   ../src/tucucore/drugmodelimport.cpp
	modified:   ../src/tucucore/drugmodelimport.h
	new file:   ../src/tucucore/iresidualerrormodel.h
	modified:   ../src/tucucore/makefile
	modified:   ../src/tucucore/residualerrormodel.h
	new file:   ../src/tucucore/residualerrormodelextractor.cpp
	new file:   ../src/tucucore/residualerrormodelextractor.h
	modified:   ../test/tucucore/drugmodels/buildimatinib.h
	modified:   ../test/tucucore/test_percentilecalculator.h