TUC-157 Better checking of Operations in DrugModelChecker
For doing this I modified TinyJS to disable the use of undeclared
variables. Instead of using a default 0 value, it raises an exception.

I added a method in JSOperation to check the operation. It removes
parts of the expression that were added in the constructor. It also
removes the last return statement.

Now the error message is more pertinent and clearly identifies
the missing variables. It also outputs the source expression with line
numbers.

	modified:   libs/tiny-js-master-20170629/TinyJS.cpp
	modified:   libs/tiny-js-master-20170629/TinyJS.h
	modified:   src/tucucore/drugmodelchecker.cpp
	modified:   src/tucucore/operation.cpp
	modified:   src/tucucore/operation.h